Add files via upload
These are my first two nuke tools i ever created, a  luminance grade knob and a AoV grade knob, these two tools make a simple to use grade node  that allow the user to pick which AoV's they want to quickly and easily adjust on a per-use case. This is useful for that quick note in dailies when you're lead/super tells you to tweak values on a specific area "just a touch".